+++ routes/api/v1/task-boards.php
@@ -10,30 +10,30 @@ use Illuminate\Support\Facades\Route;
         ->middlewareFor(['update'], 'check.permission:taskboard.boards.edit')
         ->middlewareFor(['destroy'], 'check.permission:taskboard.boards.delete');
     Route::post('boards', [TaskBoardController::class, 'store'])->middleware('check.permission:taskboard.boards.create')->name('boards.store');
-    Route::post('boards/{board}/members', [TaskBoardController::class, 'addMember']);
-    Route::delete('boards/{board}/members/{member}', [TaskBoardController::class, 'removeMember']);
-    Route::post('boards/{board}/columns', [TaskBoardController::class, 'addColumn']);
-    Route::put('boards/{board}/columns/{column}', [TaskBoardController::class, 'updateColumn']);
-    Route::delete('boards/{board}/columns/{column}', [TaskBoardController::class, 'removeColumn']);
+    Route::post('boards/{board}/members', [TaskBoardController::class, 'addMember'])->middleware('check.permission:taskboard.boards.edit');
+    Route::delete('boards/{board}/members/{member}', [TaskBoardController::class, 'removeMember'])->middleware('check.permission:taskboard.boards.edit');
+    Route::post('boards/{board}/columns', [TaskBoardController::class, 'addColumn'])->middleware('check.permission:taskboard.boards.edit');
+    Route::put('boards/{board}/columns/{column}', [TaskBoardController::class, 'updateColumn'])->middleware('check.permission:taskboard.boards.edit');
+    Route::delete('boards/{board}/columns/{column}', [TaskBoardController::class, 'removeColumn'])->middleware('check.permission:taskboard.boards.edit');
     Route::get('boards/{board}/labels', [TaskBoardController::class, 'labels']);
-    Route::post('boards/{board}/labels', [TaskBoardController::class, 'addLabel']);
+    Route::post('boards/{board}/labels', [TaskBoardController::class, 'addLabel'])->middleware('check.permission:taskboard.boards.edit');
 
     // Tasks
     Route::prefix('boards/{board}/tasks')->group(function () {
         Route::get('/', [BoardTaskController::class, 'index']);
         Route::post('/', [BoardTaskController::class, 'store'])->middleware('check.permission:taskboard.tasks.create');
         Route::get('/{task}', [BoardTaskController::class, 'show']);
-        Route::put('/{task}', [BoardTaskController::class, 'update']);
-        Route::delete('/{task}', [BoardTaskController::class, 'destroy']);
-        Route::post('/{task}/move', [BoardTaskController::class, 'move']);
-        Route::post('/{task}/assign', [BoardTaskController::class, 'assign']);
-        Route::post('/{task}/comments', [BoardTaskController::class, 'addComment']);
-        Route::post('/{task}/checklists', [BoardTaskController::class, 'addChecklist']);
-        Route::post('/{task}/time-entries', [BoardTaskController::class, 'addTimeEntry']);
-        Route::post('/{task}/attachments', [BoardTaskController::class, 'addAttachment']);
-        Route::post('/{task}/labels', [BoardTaskController::class, 'addLabel']);
-        Route::post('/{task}/watchers', [BoardTaskController::class, 'addWatcher']);
-        Route::post('/{task}/dependencies', [BoardTaskController::class, 'addDependency']);
+        Route::put('/{task}', [BoardTaskController::class, 'update'])->middleware('check.permission:taskboard.tasks.manage');
+        Route::delete('/{task}', [BoardTaskController::class, 'destroy'])->middleware('check.permission:taskboard.tasks.manage');
+        Route::post('/{task}/move', [BoardTaskController::class, 'move'])->middleware('check.permission:taskboard.tasks.manage');
+        Route::post('/{task}/assign', [BoardTaskController::class, 'assign'])->middleware('check.permission:taskboard.tasks.manage');
+        Route::post('/{task}/comments', [BoardTaskController::class, 'addComment'])->middleware('check.permission:taskboard.tasks.manage');
+        Route::post('/{task}/checklists', [BoardTaskController::class, 'addChecklist'])->middleware('check.permission:taskboard.tasks.manage');
+        Route::post('/{task}/time-entries', [BoardTaskController::class, 'addTimeEntry'])->middleware('check.permission:taskboard.tasks.manage');
+        Route::post('/{task}/attachments', [BoardTaskController::class, 'addAttachment'])->middleware('check.permission:taskboard.tasks.manage');
+        Route::post('/{task}/labels', [BoardTaskController::class, 'addLabel'])->middleware('check.permission:taskboard.tasks.manage');
+        Route::post('/{task}/watchers', [BoardTaskController::class, 'addWatcher'])->middleware('check.permission:taskboard.tasks.manage');
+        Route::post('/{task}/dependencies', [BoardTaskController::class, 'addDependency'])->middleware('check.permission:taskboard.tasks.manage');
     });
 
     // Sprints
@@ -41,11 +41,11 @@ use Illuminate\Support\Facades\Route;
         Route::get('/', [SprintController::class, 'index']);
         Route::post('/', [SprintController::class, 'store'])->middleware('check.permission:taskboard.sprints.create');
         Route::get('/{sprint}', [SprintController::class, 'show']);
-        Route::put('/{sprint}', [SprintController::class, 'update']);
-        Route::delete('/{sprint}', [SprintController::class, 'destroy']);
-        Route::post('/{sprint}/start', [SprintController::class, 'start']);
-        Route::post('/{sprint}/complete', [SprintController::class, 'complete']);
-        Route::post('/{sprint}/tasks', [SprintController::class, 'addTasks']);
-        Route::delete('/{sprint}/tasks', [SprintController::class, 'removeTasks']);
+        Route::put('/{sprint}', [SprintController::class, 'update'])->middleware('check.permission:taskboard.sprints.manage');
+        Route::delete('/{sprint}', [SprintController::class, 'destroy'])->middleware('check.permission:taskboard.sprints.manage');
+        Route::post('/{sprint}/start', [SprintController::class, 'start'])->middleware('check.permission:taskboard.sprints.manage');
+        Route::post('/{sprint}/complete', [SprintController::class, 'complete'])->middleware('check.permission:taskboard.sprints.manage');
+        Route::post('/{sprint}/tasks', [SprintController::class, 'addTasks'])->middleware('check.permission:taskboard.sprints.manage');
+        Route::delete('/{sprint}/tasks', [SprintController::class, 'removeTasks'])->middleware('check.permission:taskboard.sprints.manage');
         Route::get('/{sprint}/burndown', [SprintController::class, 'burndownChart']);
     });
